requirements.txt
django
web3
mnemonic
python-decouple
djangorestframework